requirements.txt
# ============================================
# DEPENDENCIAS - Sistema de Gestión Empresarial
# Python 3.8+
# ============================================

# Base de datos
mysql-connector-python==8.2.0

# Configuración segura
python-dotenv==1.2.1

# HTTP Requests (para APIs externas)
requests==2.31.0

# Visualización de datos
matplotlib==3.10.7
numpy==2.3.5

# Generación de PDFs
reportlab==4.4.5

# Dependencias de matplotlib
contourpy==1.3.3
cycler==0.12.1
fonttools==4.60.1
kiwisolver==1.4.9
pillow==12.0.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
six==1.17.0

# Otras dependencias
protobuf==4.21.12
certifi==2025.11.12
charset-normalizer==3.4.4
idna==3.11
urllib3==2.5.0
packaging==25.0
